Replace PureDatasource with MapBasedDatasource.pure
